service-changes

+++ controllers/App_services/appService.js
@@ -5,6 +5,30 @@ const {location}  = req.query
 
    try{
      const data = await appService.find({status :200 })
+     
+    return res.status(200).json(data)
+   }
+   catch(err){
+    return res.status(201).json({msg :"Internel server error"})
+   }
+}
+async function fetchAdminServices(req ,res){
+const {location}  = req.query
+
+   try{
+     const data = await appService.find({ })
+     
+    return res.status(200).json(data)
+   }
+   catch(err){
+    return res.status(201).json({msg :"Internel server error"})
+   }
+}
+async function fetchAdminServicesByID(req ,res){
+const {_id}  = req.query
+
+   try{
+     const data = await appService.findById(_id)
     return res.status(200).json(data)
    }
    catch(err){
@@ -13,11 +37,16 @@ const {location}  = req.query
 }
 
 async function changeStatus(req,res){
-    const status = req.body
-      await appService.updateOne(
+    const status = req.body.status
+      try{await appService.updateOne(
       { _id: req.body._id},  // Indexed _id filter
       { $set: { status: status }}
     );
+    res.status(200).json({message: `Changed Status to :${status == 200 ? 'Live' : 'Dead'} `})}
+    catch(err){
+      console.log(err)
+      res.status(201).json({message : "Internal Server error"})
+    }
 }
 
 async function addNewAppService(req,res){
@@ -42,4 +71,4 @@ async function deleteAllService(req ,res){
 
    }
 }
-module.exports = {fetchAllServices ,addNewAppService ,changeStatus , deleteAllService}
+module.exports = {fetchAllServices ,addNewAppService ,changeStatus , deleteAllService ,fetchAdminServices ,fetchAdminServicesByID}
