+++ routes/v1Routes.js
@@ -2,12 +2,13 @@ const express = require('express');
 const {fetchAllCenters, addNewCenter , suspendStation } = require("../controllers/Centers/centerController");
 const { verifyToken, verifyAdminToken } = require('../Middleware/veriyToken');
 const router = express.Router();
-const {fetchAllServices, changeStatus, addNewAppService, deleteAllService}  = require("../controllers/App_services/appService");
+const {fetchAllServices, changeStatus, addNewAppService, deleteAllService, fetchAdminServices}  = require("../controllers/App_services/appService");
 const { bookOrder, fetchOrdersbyStationID } = require('../controllers/Order/orderController');
 const { fetchUserDetails, updateUserDetail } = require('../controllers/User/userController');
 const { checkUserAuth } = require('../Middleware/checkUserAuth');
 const { verifyUserOTP } = require('../controllers/User/userAuthController');
-const { fetchAllWhyUs, addWhyUs, updateWhyUs } = require('../controllers/WhyUS/WhyUsController');
+const { fetchAllWhyUs, addWhyUs, updateWhyUs, fetchWhyUsById } = require('../controllers/WhyUS/WhyUsController');
+const { addAdmin, adminLogin } = require('../controllers/admin/adminController');
 
 router.get('/ping', (req, res) => {
   res.json({ message: 'Pong' });
@@ -16,7 +17,9 @@ router.get('/ping', (req, res) => {
 // USER TASK APIs
 // router.get('/fetch-live-station' , fetchAllCenters)
 router.get("/fetch-whyUs" , fetchAllWhyUs )
+router.get("/fetch-whyUs-ById" , fetchWhyUsById )
 router.get('/fetch-app-services' ,fetchAllServices )
+
 router.post('/book-order', verifyToken, bookOrder )
     // USER AUTHx
     router.get('/gen-otp' , ()=>{})
@@ -31,13 +34,15 @@ router.get('/fetch-user-history' ,verifyToken, checkUserAuth, ()=>{})
 router.get('/fetch-request-by-id' , verifyToken ,fetchOrdersbyStationID) 
 
 // ADMIN TASK APIs *******************
+router.post('/admin-login' ,adminLogin )
+router.post('/add-admin'  ,  addAdmin)
 router.get('/delete-all-services' , verifyAdminToken , deleteAllService)
-router.post("/change-service-status" , verifyAdminToken , changeStatus)
+router.post("/change-service-status" ,  changeStatus)
 router.post("/add-app-service" , verifyAdminToken , addNewAppService)
-
+router.get('/fetch-admin-app-services' ,fetchAdminServices )
 router.post('/add-new-station' ,verifyAdminToken ,addNewCenter )
 router.post('/suspend-station' ,verifyAdminToken ,suspendStation )
 
-router.post('/add-whyus' ,verifyAdminToken, addWhyUs)
-router.post('/update-whyus', verifyAdminToken ,updateWhyUs )
+router.post('/add-whyus' , addWhyUs)
+router.post('/update-whyus',updateWhyUs )
 module.exports = router;
